requirements.txt
langchain
langchain-openai
langgraph
atlassian-python-api
python-dotenv
pydantic